requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
pillow
tqdm
